SQL: Fix db table detection on first run.

+++ src/sql/helpers/table-exists.sql
@@ -1,4 +1,4 @@
-SELECT COUNT(*) AS exists
+SELECT COUNT(*)::INT AS exists
 FROM information_schema.tables
 WHERE table_schema = 'public'
 AND table_name = $1
